+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   Thanks, @jangaraj!
 * Add gracefulness when using the new data source health check endpoint.
   Apparently, this is not implemented thoroughly for all data source types yet.
+* Add new factory methods `GrafanaApi.{from_url(),from_env()}`.
 
 
 ## 2.3.0 (2022-05-26)